refactor: moved conditions into functions
- moves the massive if statement conditional in the script into separate functions each with their own options
- Allows the user to call multiple options through different arguments.
- This also adjusts the pylint to work with the new django settings
- Renames the file to execute.py
